added gpu processing

+++ feat_extract_charades.py
@@ -14,6 +14,10 @@ from natsort import natsorted
 import time
 from tqdm import tqdm
 
+device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+print(f"Device: {device}")
+# device = torch.device("cpu")
+
 extraCfg = edict({
     "type": "ViP",
     "temporal_size": 12,
@@ -29,7 +33,7 @@ clipconfig.vision_additional_config = extraCfg
 
 checkpoint = torch.load("/ssd_scratch/cvit/varun/pretrain_clipvip_base_32.pt")
 cleanDict = { key.replace("clipmodel.", "") : value for key, value in checkpoint.items() }
-model =  CLIPModel(config=clipconfig)
+model =  CLIPModel(config=clipconfig).to(device)
 model.load_state_dict(cleanDict)
 
 tokenizer = CLIPTokenizerFast.from_pretrained("openai/clip-vit-base-patch32")
@@ -39,7 +43,6 @@ def extract_text_features(texts, tokenizer, model):
     tokens = tokenizer(texts, padding=True, return_tensors="pt")
     textOutput = model.get_text_features(**tokens)
     return textOutput
-
 
 
 def read_video_pyav(container, indices):
@@ -53,24 +56,6 @@ def read_video_pyav(container, indices):
         if i >= start_index and i in indices:
             frames.append(frame)
     return np.stack([x.to_ndarray(format="rgb24") for x in frames])
-
-
-# def sample_frame_indices(clip_len, frame_sample_rate, seg_len):
-#     '''
-#     Sample a given number of frame indices from the video.
-#     Args:
-#         clip_len (`int`): Total number of frames to sample.
-#         frame_sample_rate (`int`): Sample every n-th frame.
-#         seg_len (`int`): Maximum allowed index of sample's last frame.
-#     Returns:
-#         indices (`List[int]`): List of sampled frame indices
-#     '''
-
-#     end_idx = seg_len
-#     start_idx = 0
-#     indices = np.linspace(start_idx, end_idx, num=clip_len)
-#     indices = np.clip(indices, start_idx, end_idx - 1).astype(np.int64)
-#     return indices
 
 def sample_frame_indices(clip_len, frame_sample_rate, seg_len):
     '''
@@ -99,7 +84,7 @@ def video_loader(video_path, processor, model):
     
     indices = sample_frame_indices(clip_len=clip_len, frame_sample_rate=fcount//clip_len, seg_len=fcount)
     video = read_video_pyav(container, indices)
-    pixel_values = processor(videos=list(video), return_tensors="pt").pixel_values
+    pixel_values = processor(videos=list(video), return_tensors="pt").pixel_values.to(device)
     B, N, C, H, W = pixel_values.shape
     vv = pixel_values.reshape(-1, C, H, W)
     inputs = {
@@ -111,35 +96,26 @@ def video_loader(video_path, processor, model):
     
     return video_features
 
-
-
 output_dir = r'/ssd_scratch/cvit/varun/charades_clipvip_feats'
 if not os.path.exists(output_dir):
     os.makedirs(output_dir)
 
-# msrvtt_annotations_path = r'/home2/varungupta/clip-vip_feat_extract/jupyter_stuff/msrvtt_annotations_formatted.json'
-# msrvtt_annotations_path = r'/home2/varungupta/clip-vip_feat_extract/mini.json'
-charades_videos_path = r'/ssd_scratch/cvit/varun/MSRVTT/videos/all'
+charades_videos_path = r'/ssd_scratch/cvit/varun/Charades_v1_480'
 
-videos_list = natsorted(glob(charades_videos_path + '/*.mp4'))[:10]
-# msrvtt_annotations = json.load(open(msrvtt_annotations_path, 'r'))
-
-
-#for each video, we create a npy file.
-# all_videos = sorted(list(msrvtt_annotations.keys()), reverse=True)
+videos_list = natsorted(glob(charades_videos_path + '/*.mp4'))
 
 for video in tqdm(videos_list):
+    st = time.time() 
     vid_name = video.split('/')[-1].split('.')[0]
+
+    if os.path.exists(os.path.join(output_dir, vid_name + '.npy')):
+        continue
+
     if os.path.exists(os.path.join(output_dir, video + '.npy')):
         continue
-    st = time.time()
-    # video_data = {}
-    # contents = msrvtt_annotations[video]
-    
-    # video_path = os.path.join(msrvtt_videos_path, video + '.mp4')
 
     vid_feat = video_loader(video, processor, model)
     vid_feat = vid_feat.detach().cpu().numpy()
-    print(vid_feat.shape)
-        
+    
     np.save(os.path.join(output_dir, vid_name + '.npy'), vid_feat)
+    print(f"Time taken for {vid_name} is {time.time() - st} seconds")
